print-list

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -1,5 +1,7 @@
-#include "lists.h"
+#include <stdlib.h>
+#include <string.h>
 #include <stdio.h>
+#include "lists.h"
 
 /**
  * print_list - Prints all the elements of a list_t list.
@@ -8,7 +10,7 @@
  * Return: The number of nodes in h.
  */
 
-size_t print_list(const list_t *h)
+size_t print_list int(const list_t *h)
 {
 	size_t nodes = 0;
